refactor(attachment): move shared access-check to service layer and avoid double query
Move resolveAccessibleAttachment from routes/attachment/ (owned by the
legacy route) to service/attachment/, matching the neutral placement
already used by attachment-snapshot.ts for logic shared by both the
apiv3 and legacy attachment routes. Also add an optional populate
argument so the apiv3 GET route resolves creator in the same query
instead of a follow-up populate() round trip.
